+++ compare-frontend/src/app/modules/wizard/wizard/step/config/config.component.ts
@@ -3,6 +3,7 @@ import { FormControl, FormGroup } from '@angular/forms';
 import { Partition, RiesgosProduct, RiesgosState, RiesgosStep, ScenarioName } from 'src/app/state/state';
 import * as AppActions from 'src/app/state/actions';
 import { Store } from '@ngrx/store';
+import { WizardComposite } from '../../../wizard.service';
 
 
 @Component({
@@ -14,24 +15,17 @@ export class ConfigComponent implements OnInit {
 
   @Input() scenario!: ScenarioName;
   @Input() partition!: Partition;
-  @Input() step!: RiesgosStep["step"];
-  @Input() products!: RiesgosProduct[];
+  @Input() data!: WizardComposite;
   public formGroup: FormGroup = new FormGroup({});
 
   constructor(private store: Store<{ riesgos: RiesgosState }>) {}
 
   ngOnInit(): void {
-    for (const input of this.step.inputs) {
-      // Kind of flaky: 
-      // We interpret any input with an options-array as user-configurable, 
-      // and all others as outputs of upstream processes. 
+    for (const input of this.data.inputs) {
       if (input.options) {
-        const id = input.id;
-        const existingValue = this.products.find(p => p.id === id)?.value;
-        const existingRef = this.products.find(p => p.id === id)?.reference;
-        const existingDefault = input.default;
+        const existingValue = input.currentValue;
         // const firstOption = input.options[0];  // when first option is set as a default value from the start, no config-action is fired, causing error when hitting "execute" without selecting another option.
-        this.formGroup.addControl(id, new FormControl(existingValue || existingRef || existingDefault || ''));
+        this.formGroup.addControl(input.label, new FormControl(existingValue || ''));
       }
     }